Update : token handler of Post routes
- Updated how token is handled, of Post-related routes
- Use HTTPBearer, instead of OAuth2PasswordBearer

+++ client-API-gateway/app/routes/posts.py
@@ -1,13 +1,13 @@
-from fastapi import APIRouter, HTTPException, Depends, Body, Request
-from fastapi.security import OAuth2PasswordBearer
-from httpx import Response
+from fastapi import APIRouter, HTTPException, Depends, Body
+from fastapi.security import HTTPBearer
 import httpx
 
 import load_envs as _envs
 
 router = APIRouter()
 
-oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")
+token_auth_scheme = HTTPBearer(auto_error=False)
+# NOTE : [oauth2_scheme] obj looks like this : { scheme = 'Bearer', credentials = 'xxxtokenxxx' }
 
 
 @router.get('/')
@@ -25,13 +25,13 @@ async def testConnect():
 
 
 @router.get('/rooms')
-async def getRooms(*, recommend = False, request: Request):
+async def getRooms(*, recommend: bool = False, token: str = Depends(token_auth_scheme)):
     try:
-        res: Response = None
-
         if recommend:
-            _bearer_and_token = request.headers.get('authorization')
-            _headers = {'Authorization': _bearer_and_token}
+            if token is None:
+                return HTTPException(status_code=401, detail="Token is required to get Reccomend Rooms")
+                
+            _headers = {'Authorization': f"Bearer {token.credentials}"}
             res = httpx.get(f"{_envs.POST_SERVICES_URL}/posts/rooms?recommend={recommend}", headers=_headers)
         else:
             res = httpx.get(f"{_envs.POST_SERVICES_URL}/posts/rooms")
@@ -46,11 +46,10 @@ async def getRooms(*, recommend = False, request: Request):
 
 
 @router.get('/rooms/{room_id}')
-async def getById(room_id: int, request: Request):
+async def getById(room_id: int, token: str = Depends(token_auth_scheme)):
     try:
-        _bearer_and_token = request.headers.get('authorization')
-        if _bearer_and_token is not None:
-            _headers = {'Authorization': _bearer_and_token}
+        if token is not None:
+            _headers = {'Authorization': f"Bearer {token.credentials}"}
             res = httpx.get(f"{_envs.POST_SERVICES_URL}/posts/rooms/{room_id}", headers=_headers)
         else:
             res = httpx.get(f"{_envs.POST_SERVICES_URL}/posts/rooms/{room_id}")
@@ -64,9 +63,12 @@ async def getById(room_id: int, request: Request):
 
 
 @router.post('/rooms')
-async def createRoom(payload: dict = Body(), token: str = Depends(oauth2_scheme)):
+async def createRoom(payload: dict = Body(), token: str = Depends(token_auth_scheme)):
     try:
-        _headers = {'Authorization': 'Bearer ' + token}
+        if token is None:
+            return HTTPException(status_code=401, detail="Token is required")
+
+        _headers = {'Authorization': f"Bearer {token.credentials}"}
         res = httpx.post(f"{_envs.POST_SERVICES_URL}/posts/rooms", json=payload, headers=_headers)
         _res = res.json()
         if res.status_code != 200:
@@ -77,9 +79,12 @@ async def createRoom(payload: dict = Body(), token: str = Depends(oauth2_scheme)
 
 
 @router.patch('/rooms/{room_id}')
-async def updateRoom(room_id: int, payload: dict = Body(), token: str = Depends(oauth2_scheme)):
+async def updateRoom(room_id: int, payload: dict = Body(), token: str = Depends(token_auth_scheme)):
     try:
-        _headers = {'Authorization': 'Bearer ' + token}
+        if token is None:
+            return HTTPException(status_code=401, detail="Token is required")
+
+        _headers = {'Authorization': f"Bearer {token.credentials}"}
         res = httpx.patch(f"{_envs.POST_SERVICES_URL}/posts/rooms/{room_id}", json=payload, headers=_headers)
         _res = res.json()
         if res.status_code != 200:
@@ -90,9 +95,12 @@ async def updateRoom(room_id: int, payload: dict = Body(), token: str = Depends(
 
 
 @router.delete('/room/{room_id}')
-async def deleteRoom(room_id: int, token: str = Depends(oauth2_scheme)):
+async def deleteRoom(room_id: int, token: str = Depends(token_auth_scheme)):
     try:
-        _headers = {'Authorization': 'Bearer ' + token}
+        if token is None:
+            return HTTPException(status_code=401, detail="Token is required")
+
+        _headers = {'Authorization': f"Bearer {token.credentials}"}
         res = httpx.delete(f"{_envs.POST_SERVICES_URL}/posts/rooms/{room_id}", headers=_headers)
         _res = res.json()
         if res.status_code != 200:
@@ -103,9 +111,12 @@ async def deleteRoom(room_id: int, token: str = Depends(oauth2_scheme)):
 
 
 @router.post('/rooms/{room_id}/add-message')
-async def addMessage(room_id: int, payload: dict = Body(), token: str = Depends(oauth2_scheme)):
+async def addMessage(room_id: int, payload: dict = Body(), token: str = Depends(token_auth_scheme)):
     try:
-        _headers = {'Authorization': 'Bearer ' + token}
+        if token is None:
+            return HTTPException(status_code=401, detail="Token is required")
+
+        _headers = {'Authorization': f"Bearer {token.credentials}"}
         res = httpx.post(f"{_envs.POST_SERVICES_URL}/posts/rooms/{room_id}/add-message", json=payload, headers=_headers)
         _res = res.json()
         if res.status_code != 200:
